compute/requirements.txt
fastapi
mangum
pyjwt